locust-tasks/requirements.txt
Click==7.0
Cython==0.29.10
Flask==1.0.3
Jinja2==2.10.1
MarkupSafe==1.1.1
Werkzeug==0.15.4
certifi==2019.3.9
chardet==3.0.4
gevent==1.4.0
greenlet==0.4.15
idna==2.8
itsdangerous==1.1.0
locustio==0.11.0
msgpack==0.6.1
msgpack-python==0.5.6
pyzmq==18.0.1
requests==2.22.0
six==1.12.0
urllib3==1.25.3
virtualenv==1.11.6
wsgiref==0.1.2
